+++ noodles-bcf/CHANGELOG.md
@@ -7,6 +7,11 @@
   * bcf: Set minimum supported Rust version (MSRV) to 1.56.0 in package
     manifest.
 
+### Fixed
+
+  * bcf/writer/vcf_record/site: Avoid overflow when the alternate allele count
+    is `usize::MAX`.
+
 ## 0.12.0 - 2022-01-27
 
 ### Added